funciona la notificacion
joderrrr

+++ app/src/main/java/com/example/foodtracker/NotificationData.kt
@@ -1,4 +1,4 @@
 package com.example.foodtracker
 
 // Define a data class to hold notification data
-data class NotificationData(val id: Int, val title: String, val content: String, val dateTime: Long)
+data class NotificationData(val id: Long, val title: String, val content: String, val dateTime: Long)
